podcast-gen/pyproject.toml
[project]
name = "podcast-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.63.0",
    "pillow>=12.1.1",
]
